Move component scoped types and consts to component folder
